refactor(account): add everything account related inside charts/cards/crud sub-folders
