+++ FinEdgeFrontend/src/Components/AccountsPage/Cards/AccountQuickStatsCardHighestBalance.tsx
